+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added exponential backoff window module
 - Added parametric Galois LFSR module with optional whitening feature
+- Added `cf_math_pkg`: Constant Function implementations of mathematical functions for HDL elaboration
 
 ### Changed
 - Parametric payload data type for `rr_arb_tree`